Initial upload
Hardware is almost done: PCB is done, first version of the case is done but needs improvements for at least dust protection. Software needs to be written, there are just some example codes found on the internet for simple calculators that can be used as the base but needs to be fully developed.